+++ src/js/app/reducers/index.js
@@ -1,13 +1,16 @@
 // @flow
 import { combineReducers } from 'redux';
-import { routerReducer as router } from 'react-router-redux/reducer';
+import { connectRouter } from 'connected-react-router';
 import layout from './layoutReducer';
 import todoManager from './todoManagerReducer';
 import chuckNorris from './chuckNorrisReducer';
 
-export default combineReducers({
-  router,
-  layout,
-  todoManager,
-  chuckNorris,
-});
+const reducers = (history: *) =>
+  combineReducers({
+    router: connectRouter(history),
+    layout,
+    todoManager,
+    chuckNorris,
+  });
+
+export default reducers;