+++ src/js/app/store/index.js
@@ -4,7 +4,7 @@
  */
 import type { GenericStoreEnchancer } from 'redux';
 import { applyMiddleware, compose, createStore, StoreCreator } from 'redux';
-import routerMiddleware from 'react-router-redux/middleware';
+import { routerMiddleware } from 'connected-react-router';
 import { createEpicMiddleware } from 'redux-observable';
 import reduxDevToolsExtension from './reduxDevtoolsExtension';
 import rootEpic from '../epics/index';
@@ -14,8 +14,6 @@ import env from '../utils/env';
 const configureStore = (history: *): StoreCreator => {
   const epicMiddleware = createEpicMiddleware();
 
-  // To allow epic to change use `push(..)` and such to change the routing.
-  // See https://github.com/reactjs/react-router-redux#pushlocation-replacelocation-gonumber-goback-goforward
   const routerHistoryMiddleware = routerMiddleware(history);
 
   let enhancer: GenericStoreEnchancer = applyMiddleware(epicMiddleware, routerHistoryMiddleware);
@@ -27,7 +25,7 @@ const configureStore = (history: *): StoreCreator => {
     );
   }
 
-  const store = createStore(reducers, enhancer);
+  const store = createStore(reducers(history), enhancer);
 
   epicMiddleware.run(rootEpic);