+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.15.0 - 2018-12-28
+
+- Replaced `react-router-redux` with `connected-react-router` because the earlier won't work with Redux v6.
+- Run Prettier on Git pre commit.
+- Removed `uglifyjs-webpack-plugin` to use Webpack's built-in `terser-webpack-plugin`.
+
 ## 0.14.0 - 2018-10-04
 
 - Babel - Upgrade to v7.